runner - work around Java infinite print bug by restarting container
When you do a Java infinite print program, unlike Python or Ruby,
the container is unable to run Java programs any more. strace wasn't
revealing in debugging this. Will try to revisit this but in the
mean time if any language exits unsuccessfully we'll always restart
the container. This adds 10 seconds of latency.